Cleaner error serialization and instantiation
Previously, serializerr was used to serialize Errors before they were
returned to clients.  Instead, now we take advantage of the fact
that KasockiError has a toJSON method that can explicitly
prepare and object for serialization, and also of bunyan's
configurable serializers to use that toJSON method when
logging.

serializerr worked great, but not for logging.  We were going
to need a custom bunyan err serializer to have extra error properties
logged.  By implementing toJSON on the KasockiError class,
we solve both logging and socket.io serialization problems.